pyproject.toml
[tool.poetry]
name = "emkademy"
version = "0.1.0"
description = ""
authors = ["CanerOren <canerorenn@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
pydantic = "^2"
hydra-core = "^1.3"
pandas = "~=2.3"
nltk = "~=3.9" 
symspellpy = "~=6.9"
fsspec = {version = "~=2025.10", extras = ["gcs"]}
gcsfs = "~=2025.10"
google-cloud-secret-manager = "~=2.25"
fastparquet = "~=2024.11"
dvc = {version = "~=3.63", extras = ["gdrive", "gs"]}
dask = {version = "~=2025.10", extras = ["complete"]}
distributed = "~=2025.10"
dask-cloudprovider = "~=2025.9"
dask-ml = "~=2025.1"
bokeh = "~=3.8"
pyOpenSSL = ">=23.1.1"

tokenizers = "~=0.22"
transformers = "~=4.57.1"


[tool.poetry.group.dev.dependencies]
jupyterlab = "~=4.4"
pytest = "~7.3"
black = "*"
isort = "*"
flake8 = "*"
mypy = "*"

types-PyYAML = "~=6.0.12.20"
types-decorator = "~=5.2.0.20"
types-paramiko = "~=4.0.0.20"
types-protobuf = "~=6.32.1.20"
types-pyOpenSSL = "~=24.1.0.20"
types-python-dateutil = "~=2.9.0.20"
types-requests = "~=2.32.4.20"
types-setuptools = "~=80.9.0.20"
types-six = "~=1.17.0.20"
types-typed-ast = "~=1.5.8.7"
types-urllib3 = "~=1.26.25.14"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=120

[tool.isort]
profile="black"
multi_line_output= 3 
include_trailing_comma=true
force_grid_wrap= 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party= ["cybulde"]
default_section = "THIRDPARTY"
sections= ["FUTURE","STDLIB","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]

[tool.mypy]
python_version = "3.12"
show_error_codes = true
ignore_missing_imports = true
install_types= true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub= true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
